extract db
